Refactor flight detail map view calculation
